refactor: splitt data handler into multiple files
